requirements.txt
Flask==2.1.0
mysql-connector==2.2.9
tensorflow==2.6.0
